masternode-payments.cpp/h
