+++ src/darksend.cpp
@@ -1407,25 +1407,28 @@ bool CDarksendPool::DoAutomaticDenominating(bool fDryRun, bool ready)
 
     // ** find the coins we'll use
     std::vector<CTxIn> vCoins;
-    int64_t nValueMin = CENT;
-    int64_t nValueIn = 0;
+    CAmount nValueMin = CENT;
+    CAmount nValueIn = 0;
+
+    CAmount nOnlyDenominatedBalance;
+    CAmount nBalanceNeedsDenominated;
 
     // should not be less than fees in DARKSEND_COLLATERAL + few (lets say 5) smallest denoms
-    int64_t nLowestDenom = DARKSEND_COLLATERAL + darkSendDenominations[darkSendDenominations.size() - 1]*5;
+    CAmount nLowestDenom = DARKSEND_COLLATERAL + darkSendDenominations[darkSendDenominations.size() - 1]*5;
 
     // if there are no DS collateral inputs yet
     if(!pwalletMain->HasCollateralInputs())
         // should have some additional amount for them
         nLowestDenom += DARKSEND_COLLATERAL*4;
 
-    int64_t nBalanceNeedsAnonymized = nAnonymizeTransferAmount*COIN - pwalletMain->GetAnonymizedBalance();
+    CAmount nBalanceNeedsAnonymized = nAnonymizeTransferAmount*COIN - pwalletMain->GetAnonymizedBalance();
 
     // if balanceNeedsAnonymized is more than pool max, take the pool max
     if(nBalanceNeedsAnonymized > DARKSEND_POOL_MAX) nBalanceNeedsAnonymized = DARKSEND_POOL_MAX;
 
     // if balanceNeedsAnonymized is more than non-anonymized, take non-anonymized
-    int64_t nBalanceNotYetAnonymized = pwalletMain->GetBalance() - pwalletMain->GetAnonymizedBalance();
-    if(nBalanceNeedsAnonymized > nBalanceNotYetAnonymized) nBalanceNeedsAnonymized = nBalanceNotYetAnonymized;
+    CAmount nAnonymizableBalance = pwalletMain->GetAnonymizableBalance();
+    if(nBalanceNeedsAnonymized > nAnonymizableBalance) nBalanceNeedsAnonymized = nAnonymizableBalance;
 
     if(nBalanceNeedsAnonymized < nLowestDenom)
     {
@@ -1444,7 +1447,13 @@ bool CDarksendPool::DoAutomaticDenominating(bool fDryRun, bool ready)
 
         if (pwalletMain->SelectCoinsDark(nValueMin, 9999999*COIN, vCoins, nValueIn, -2, 0))
         {
-            if(!fDryRun) return CreateDenominated(nBalanceNeedsAnonymized);
+            nOnlyDenominatedBalance = pwalletMain->GetDenominatedBalance(true, false, false);
+            nBalanceNeedsDenominated = nBalanceNeedsAnonymized - nOnlyDenominatedBalance;
+
+            if(nBalanceNeedsDenominated > nValueIn) nBalanceNeedsDenominated = nValueIn;
+
+            if(!fDryRun) return CreateDenominated(nBalanceNeedsDenominated);            
+
             return true;
         } else {
             LogPrintf("DoAutomaticDenominating : Can't denominate - no compatible inputs left\n");
@@ -1453,6 +1462,11 @@ bool CDarksendPool::DoAutomaticDenominating(bool fDryRun, bool ready)
         }
 
     }
+
+    nOnlyDenominatedBalance = pwalletMain->GetDenominatedBalance(true, false, false);
+    nBalanceNeedsDenominated = nBalanceNeedsAnonymized - nOnlyDenominatedBalance;
+
+    if(!fDryRun && nBalanceNeedsDenominated > nOnlyDenominatedBalance) return CreateDenominated(nBalanceNeedsDenominated);
 
     //check to see if we have the collateral sized inputs, it requires these
     if(!pwalletMain->HasCollateralInputs()){
@@ -1467,14 +1481,6 @@ bool CDarksendPool::DoAutomaticDenominating(bool fDryRun, bool ready)
     if(!sessionFoundMasternode){
         int nUseQueue = rand()%100;
         UpdateState(POOL_STATUS_ACCEPTING_ENTRIES);
-
-        sessionTotalValue = pwalletMain->GetTotalValue(vCoins);
-
-        //randomize the amounts we mix
-        if(sessionTotalValue > nBalanceNeedsAnonymized) sessionTotalValue = nBalanceNeedsAnonymized;
-
-        double fTransferSubmitted = (sessionTotalValue / COIN);
-        LogPrintf("Submitting Darksend for %f TX - sessionTotalValue %d\n", fTransferSubmitted, sessionTotalValue);
 
         if(pwalletMain->GetDenominatedBalance(true, true) > 0){ //get denominated unconfirmed inputs
             LogPrintf("DoAutomaticDenominating -- Found unconfirmed denominated outputs, will wait till they confirm to continue.\n");
@@ -1526,8 +1532,8 @@ bool CDarksendPool::DoAutomaticDenominating(bool fDryRun, bool ready)
                 // connect to Masternode and submit the queue request
                 if(ConnectNode((CAddress)addr, NULL, true)){
 
-                    CNode* pNode = FindNode(addr);
-                    if(pNode)
+                    CNode* pnode = FindNode(addr);
+                    if(pnode)
                     {
                         std::string strReason;
                         if(txCollateral == CTransaction()){
@@ -1547,8 +1553,8 @@ bool CDarksendPool::DoAutomaticDenominating(bool fDryRun, bool ready)
                         vecMasternodesUsed.push_back(dsq.vin);
                         sessionDenom = dsq.nDenom;
 
-                        pNode->PushMessage("dsa", sessionDenom, txCollateral);
-                        LogPrintf("DoAutomaticDenominating --- connected (from queue), sending dsa for %d %d - %s\n", sessionDenom, GetDenominationsByAmount(sessionTotalValue), pNode->addr.ToString().c_str());
+                        pnode->PushMessage("dsa", sessionDenom, txCollateral);
+                        LogPrintf("DoAutomaticDenominating --- connected (from queue), sending dsa for %d - %s\n", sessionDenom, pnode->addr.ToString().c_str());
                         strAutoDenomResult = "";
                         dsq.time = 0; //remove node
                         return true;
@@ -1611,13 +1617,14 @@ bool CDarksendPool::DoAutomaticDenominating(bool fDryRun, bool ready)
                     pSubmittedToMasternode = pmn;
                     vecMasternodesUsed.push_back(pmn->vin);
 
-                    std::vector<int64_t> vecAmounts;
+                    std::vector<CAmount> vecAmounts;
                     pwalletMain->ConvertList(vCoins, vecAmounts);
-                    sessionDenom = GetDenominationsByAmounts(vecAmounts);
+                    // try to get random denoms out of vecAmounts
+                    while(sessionDenom == 0)
+                        sessionDenom = GetDenominationsByAmounts(vecAmounts, true);
 
                     pnode->PushMessage("dsa", sessionDenom, txCollateral);
-                    LogPrintf("DoAutomaticDenominating --- connected, sending dsa for %d - denom %d\n", sessionDenom, GetDenominationsByAmount(sessionTotalValue));
-                    strAutoDenomResult = "";
+                    LogPrintf("DoAutomaticDenominating --- connected, sending ssa for %d\n", sessionDenom); 
                     return true;
                 }
             } else {
@@ -1921,7 +1928,7 @@ int CDarksendPool::GetDenominations(const std::vector<CTxDSOut>& vout){
 }
 
 // return a bitshifted integer representing the denominations in this list
-int CDarksendPool::GetDenominations(const std::vector<CTxOut>& vout){
+int CDarksendPool::GetDenominations(const std::vector<CTxOut>& vout, bool fRandDenom){
     std::vector<pair<int64_t, int> > denomUsed;
 
     // make a list of denominations, with zero uses
@@ -1945,7 +1952,7 @@ int CDarksendPool::GetDenominations(const std::vector<CTxOut>& vout){
     // if the denomination is used, shift the bit on.
     // then move to the next
     BOOST_FOREACH (PAIRTYPE(int64_t, int)& s, denomUsed)
-        denom |= s.second << c++;
+        denom |= ((fRandDenom ? rand()%2 : 1) * s.second) << c++;
 
     // Function returns as follows:
     //
@@ -1958,7 +1965,7 @@ int CDarksendPool::GetDenominations(const std::vector<CTxOut>& vout){
 }
 
 
-int CDarksendPool::GetDenominationsByAmounts(std::vector<int64_t>& vecAmount){
+int CDarksendPool::GetDenominationsByAmounts(std::vector<int64_t>& vecAmount, bool fRandDenom){
     CScript e = CScript();
     std::vector<CTxOut> vout1;
 
@@ -1971,7 +1978,7 @@ int CDarksendPool::GetDenominationsByAmounts(std::vector<int64_t>& vecAmount){
         nOutputs++;
     }
 
-    return GetDenominations(vout1);
+    return GetDenominations(vout1, fRandDenom);
 }
 
 int CDarksendPool::GetDenominationsByAmount(int64_t nAmount, int nDenomTarget){
@@ -2004,12 +2011,6 @@ int CDarksendPool::GetDenominationsByAmount(int64_t nAmount, int nDenomTarget){
             nOutputs++;
         }
         LogPrintf("GetDenominationsByAmount --- %d nOutputs %d\n", v, nOutputs);
-    }
-
-    //add non-denom left overs as change
-    if(nValueLeft > 0){
-        CTxOut o(nValueLeft, e);
-        vout1.push_back(o);
     }
 
     return GetDenominations(vout1);
