+++ src/makefile.unix
@@ -148,6 +148,7 @@ OBJS= \
     obj/masternodeconfig.o \
     obj/masternodeman.o \
     obj/masternode.o \
+    obj/masternode-payments.o \
     obj/masternode-pos.o \
     obj/rpcdarksend.o \
     obj/spork.o \
