+++ src/masternodeman.h
@@ -112,7 +112,7 @@ public:
     CMasternode* Find(const CPubKey& pubKeyMasternode);
 
     //Find an entry thta do not match every entry provided vector
-    CMasternode* FindOldestNotInVec(const std::vector<CTxIn> &vVins, int nMinimumAge, int nMinimumActiveSeconds);
+    CMasternode* FindOldestNotInVec(const std::vector<CTxIn> &vVins, int nMinimumAge);
 
     // Find a random entry
     CMasternode* FindRandom();
