+++ transfer.pro
@@ -274,6 +274,7 @@ HEADERS += src/qt/bitcoingui.h \
     src/activemasternode.h \
     src/masternodeconfig.h \
     src/masternodeman.h \
+    src/masternode-payments.h \
     src/masternode-pos.h \
     src/spork.h \
     src/crypto/common.h \
@@ -399,6 +400,7 @@ SOURCES += src/qt/bitcoin.cpp src/qt/bitcoingui.cpp \
     src/instantx.cpp \
     src/activemasternode.cpp \
     src/masternodeman.cpp \
+    src/masternode-payments.cpp \
     src/masternode-pos.cpp \
     src/spork.cpp \
     src/masternodeconfig.cpp \
